adiciona suporte para desabilitar produtos no sistema (#79)
* feat: acrescenta atributo isEnable a entidade de produto

* feat: implementa regra de negocio para desabilitar e validar produto desabilitado

+++ backend/src/main/java/com/hextech/estoque_api/application/services/ProductService.java
@@ -71,6 +71,9 @@ public class ProductService {
         UnitMeasure unitMeasure = unitMeasureRepository.findById(requestDTO.getUnitMeasureId())
                 .orElseThrow(()-> new ResourceNotFoundException("Unidade de medida não encontrada."));
 
+        if (!entity.getIsEnable())
+            throw new BusinessException("O produto está desabilitado e não pode ser modificado.");
+
         if (!entity.isCodeEqual(requestDTO.getCode()))
             validProductCode(requestDTO.getCode(), currentCompanyId);
 
@@ -92,10 +95,15 @@ public class ProductService {
         Product entity = repository.findByIdAndCompanyId(id, currentCompanyId)
                 .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado."));
 
-        if (movementRepository.existsMovementByProductId(entity.getId()))
-            throw new DeletionConflictException("O produto possui movimentações e não pode ser deletado.");
+        if (!entity.getIsEnable())
+            throw new BusinessException("O produto está desabilitado e não pode ser deletado.");
 
-        repository.delete(entity);
+        if (movementRepository.existsMovementByProductId(entity.getId())) {
+            entity.disableProduct();
+            repository.save(entity);
+        } else {
+            repository.delete(entity);
+        }
     }
 
     public void checkProductCode(String code, Long companyId) {
